Initial commit
This is an initial commit of the simulation software re-thought to only use p5.js to render the molecule glyphs and handle the collision logic. The code currently only renders an array of one type of molecule on screen and assumes perfectly elastic collisions in zero gravity. The next iteration will be attempt to apply molecular dynamics.