Code cleanup + Kinetic Energy
Cleaned up functions that are not being used anymore in the Box2d implementation of the cody. Also began to write a Kinetic Energy function. The code currently only grabs the translational kinetic energy of the particles. To be implemented still is the calculation of rotational KE to add into the translational term.