Initial commit of Lennard-Jones potential
Attempted to calculate the van der Waals interactions between particles. There is something wrong with the math as well as the magnitude of the forces. The function is highly asymptotic, so the forces being generated are very large. This is causing bizarre behavior that will need to be more closely inspected.

The force function is calculated as the negative gradient of the potential field. 12*a/r^13 - 6*b/r^7, where both a and b are scaling constants.